feat: add piece manager
